// lib/types.ts
export interface HomePageContent {
    body_h2_1: string; // "Calculators by category"
    body_h2_2: string; // "Popular Calculators"
    body_h2_3: string; // "All Calculators"
    body_h2_4: string; // "Promo Title"
    body_pop_up_text: string; // "Promo Text"
    body_get_pop_up_button: string; // "Button Text"
    body_h2_5: string; // "About Title"
    body_about_us_text: string; // "About Text"
    // Ссылки для баннеров
    ad_link_1?: string; // Ссылка для google-ads-1
    ad_link_2?: string; // Ссылка для google-ads-2
    ad_link_3?: string; // Ссылка для own-ads
    ad_link_4?: string; // Ссылка для google-ads-3 (300x600)
}